fwts_mmap: fwts_page_size fix comparing size_t against -1
sysconf returns a long and we need to compare this to -1
rather than a size_t (which is always positive).

Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>

+++ src/lib/src/fwts_mmap.c
@@ -33,12 +33,12 @@
  */
 size_t fwts_page_size(void)
 {
-	size_t page_size;
+	long page_size;
 
 	page_size = sysconf(_SC_PAGESIZE);
 
 	/* If sysconf() returns -1, default it 4K */
-	return page_size == -1 ? FWTS_DEFAULT_PAGE_SIZE : page_size;
+	return (size_t)(page_size == -1 ? FWTS_DEFAULT_PAGE_SIZE : page_size);
 }
 
 /*
